requirements.txt
numpy==2.2.3
pandas==2.2.3
pytz==2025.1
mysql-connector-python==9.2.0
python-dotenv==1.0.1